Implement the deferred sending of signals.
This change implements the deferred sending of signals. All calls to
send_sig and cause_sig were replaced by send_sig_deferred and
cause_sig_deferred respectively.
All signals are buffered until reaching switch_to_user. By doing so, we
can schedule the signals without acquiring the locks on the processes
involved in the signal's delivery. Thus we don't have to do tricks by
releasing and re-acquiring process locks, but rather wait to be done
with the current kernel op and then send all the signal when we are
outside the kernel op.