Fix issues with NMI communication.
Two problem could occur:
	* First, an NMI could arrive in the middle of smp_sched_handler
(if the latter was called by context_stop or smp_schedule_sync). When
that happened, the request would be processed twice, and an assert would
fail).
	* Second, the flag value was reset even when there was no
request. That happened most of the time because context_stop would call
smp_sched_handler. Because of this we could have a race condition if
another cpu was trying to send a request.

The fix is to simply disallow calling smp_sched_handler outside the nmi
handler.